Refactored Tables and Profiles for new backend.
