+++ src/components/AdminDash/Navigation/Navigation.js
@@ -13,6 +13,44 @@ import {
 import { people, map, dashboard, pulse, userShield } from './icons'
 import Container from '../UserOptions/OptionContainer'
 
+const routes = [
+  {
+    title: 'Dashboard',
+    icon: dashboard,
+    alt: 'Icon of a dashboard',
+    path: '/dashboard',
+  },
+  {
+    title: 'Database',
+    icon: people,
+    alt: 'Icon of two people',
+    path: '/database',
+  },
+  {
+    title: 'Map',
+    icon: map,
+    alt: 'Icon of a map',
+    path: '/map',
+  },
+]
+
+const UtilitiesContainer = ({ setOpen, open }) => {
+  return (
+    <Utilities>
+      <PulseContainer>
+        <img src={pulse} alt='Icon of a pulse' />
+      </PulseContainer>
+      <IconContainer
+        onMouseEnter={() => setOpen(true)}
+        onMouseLeave={() => setOpen(false)}
+      >
+        <img src={userShield} alt='Icon of a shield with a person' />
+        {open ? <Container /> : null}
+      </IconContainer>
+    </Utilities>
+  )
+}
+
 export default function Navigation() {
   const [open, setOpen] = useState(false)
   const { pathname } = useLocation()
@@ -24,34 +62,17 @@ export default function Navigation() {
         <LogoImg src={require('../../../images/logo.png')} />
       </Logo>
       <Nav>
-        <Tab
-          onClick={() => push('/dashboard')}
-          active={pathname === '/dashboard'}
-        >
-          <img src={dashboard} alt='Icon of a dashboard' />
-          <p>Dashboard</p>
-        </Tab>
-        <Tab active={pathname === '/data'}>
-          <img src={people} alt='Icon of two people' />
-          <p>Database</p>
-        </Tab>
-        <Tab active={pathname === '/map'}>
-          <img src={map} alt='Icon of a map' />
-          <p>Map</p>
-        </Tab>
+        {routes.map(route => (
+          <Tab
+            onClick={() => push(route.path)}
+            active={pathname === route.path}
+          >
+            <img src={route.icon} alt={route.alt} />
+            <p>{route.title}</p>
+          </Tab>
+        ))}
       </Nav>
-      <Utilities>
-        <PulseContainer>
-          <img src={pulse} alt='Icon of a pulse' />
-        </PulseContainer>
-        <IconContainer
-          onMouseEnter={() => setOpen(true)}
-          onMouseLeave={() => setOpen(false)}
-        >
-          <img src={userShield} alt='Icon of a shield with a person' />
-          {open ? <Container /> : null}
-        </IconContainer>
-      </Utilities>
+      <UtilitiesContainer open={open} setOpen={setOpen} />
     </NavBar>
   )
 }
