+++ src/components/AdminDash/Table/MemberTable.test.js
@@ -1,4 +1,13 @@
-import DataMap from './DataMap'
+import React from 'react'
+import { renderWithReduxAndRouter as render } from '../../../utils/testHelpers'
+import MemberTable from './MemberTable'
+
+const neighbors = []
+const families = []
+
+test('renders MemberTable without crashing', () => {
+  render(<MemberTable userType='neighbors' Data={neighbors} Data2={families} />)
+})
 
 const data = {
   Data: [{ first_name: 'bob', last_name: 'smith' }, {}],
@@ -21,4 +30,10 @@ test('test that match is coming through as none', () => {
   const expected = 'none'
   const actual = DataMap(data)[0].match
   expect(expected).toEqual(actual)
+})
+
+test('renders MemberTable without crashing', () => {
+  const neighbors = []
+  const props = {}
+  render(<MemberTable columns={columns} data={neighbors} props={props} />)
 })
