--- src/components/AdminDash/Table/TableTitle.js
@@ -1,10 +1,0 @@
-import React from 'react'
-import { Title, TitleCtn } from './TitleStyle'
-
-export default function TableHeader(props) {
-  return (
-    <TitleCtn>
-      <Title>User Database</Title>
-    </TitleCtn>
-  )
-}
